Seagull OK

+++ include/functions.h
@@ -1647,7 +1647,7 @@ void Actor_ProcessInitChain(Actor* actor, InitChainEntry* ichain);
 f32 Math_SmoothStepToF(f32* pValue, f32 target, f32 fraction, f32 step, f32 minStep);
 void Math_ApproachF(f32* pValue, f32 target, f32 scale, f32 maxStep);
 void Math_ApproachZeroF(f32* pValue, f32 scale, f32 maxStep);
-s32 Math_SmoothStepToS(s16* pValue, s16 target, s16 scale, s16 step, s16 minStep);
+s16 Math_SmoothStepToS(s16* pValue, s16 target, s16 scale, s16 step, s16 minStep);
 void Math_ApproachS(s16* pValue, s16 target, s16 scale, s16 maxStep);
 void Color_RGBA8_Copy(Color_RGBA8* dst, Color_RGBA8* src);
 void func_801000A4(u16 sfxId);
